Lab 9/GUI/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-classic
llama-cpp-python
sentence_transformers
pypdf2
faiss-cpu
python-dotenv
streamlit